backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "cryptography>=46.0.5",
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "indpy-core>=0.1.7",
    "numpy>=2.4.3",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pymupdf>=1.27.2",
    "pytesseract>=0.3.13",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
